backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.2.1
pymongo==4.15.4
google-generativeai==0.3.0
requests==2.32.5
beautifulsoup4==4.14.3
